Updated EncounterType
Since the discharge disposition is more associated with discharging from in Inpatient visit, I modified the encounter type and updated the narrative text accordingly. I purposely made the original text different than the displayName to show that the human readable concept does not need to be exactly the same as the displayName. It just needs to be dereferenced through the originalText element of the coded concept so the association of content/meaning to the coded concept can be preserved.